add new test file; refactor scene graph; currently broken, staging for now

+++ src/geometry/geometry.cpp
@@ -23,7 +23,7 @@ Geometry make_quad(float size) {
       {1.0f, 0.0f},
   };
 
-  std::vector<uint16_t> indices = {0, 1, 2, 0, 2, 3};
+  std::vector<uint32_t> indices = {0, 1, 2, 0, 2, 3};
 
   std::vector<Vertex> vertices(positions.size());
   for (int i = 0; i < positions.size(); i++) {
@@ -95,7 +95,7 @@ Geometry make_cube(float size) {
       // Top
       vec2(0.0f, 0.0f), vec2(1.0f, 0.0f), vec2(1.0f, 1.0f), vec2(0.0f, 1.0f)};
 
-  std::vector<uint16_t> el = {0,  1,  2,  0,  2,  3,  4,  5,  6,  4,  6,  7,
+  std::vector<uint32_t> el = {0,  1,  2,  0,  2,  3,  4,  5,  6,  4,  6,  7,
                               8,  9,  10, 8,  10, 11, 12, 13, 14, 12, 14, 15,
                               16, 17, 18, 16, 18, 19, 20, 21, 22, 20, 22, 23};
