+++ src/resource/geometry_buffer.cpp
@@ -1,5 +1,6 @@
 #include <resource/geometry_buffer.h>
 #include <spdlog/spdlog.h>
+#include <stdint.h>
 GeometryBuffer::GeometryBuffer(const std::vector<Vertex> &vertices)
     : Resource(ResourceType::Geometry) {
   m_has_indices = false;
@@ -13,14 +14,14 @@ GeometryBuffer::GeometryBuffer(const std::vector<Vertex> &vertices)
 }
 
 GeometryBuffer::GeometryBuffer(const std::vector<Vertex> &vertices,
-                               const std::vector<uint16_t> &indices)
+                               const std::vector<uint32_t> &indices)
     : Resource(ResourceType::Geometry) {
   m_has_indices = true;
   glCreateBuffers(1, &vbo);
   glNamedBufferStorage(vbo, vertices.size() * sizeof(Vertex), vertices.data(),
                        0);
   glCreateBuffers(1, &ebo);
-  glNamedBufferStorage(ebo, indices.size() * sizeof(uint16_t), indices.data(),
+  glNamedBufferStorage(ebo, indices.size() * sizeof(uint32_t), indices.data(),
                        0);
   glCreateVertexArrays(1, &vao);
   glVertexArrayVertexBuffer(vao, 0, vbo, 0, sizeof(Vertex));
